build: Switch to using Zephyr modules to include the repo
In order to include the repository in the Zephyr build, we used to rely
on a custom hook in the zephyr repository's root CMakeLists.txt and
Kconfig.zephyr files, which included the nrf/ repository and this in
turn included the present nrfxlib/ one. Now that Zephyr natively supports
external modules, use this mechanism to include nrfxlib in the Zephyr build
without relying on the nrf repository.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>