bluetooth: controller: New release of ble_controller library
This version of ble_controller library does not include MPSL.
Adds select MPSL in ble_controller Kconfig as MPSL is now a
separate library that must be linked in.
Removes several APIs from ble_controller as they are replaced by
APIs in MPSL.
Updates documentation to align it with the separation of MPSL and
ble_controller.

Signed-off-by: Solveig Fure <Solveig.Fure@nordicsemi.no>