ble_controller: Add support for BT_LL_NRFXLIB_VS_INCLUDE
When the nRF BLE Controller is flashed on a separate core (e.g., with
hci_rpmsg on nRF53), the application might still need access to the
vendor specific command and event definitions.

Specififying CONFIG_BT_LL_NRFXLIB_VS_INCLUDE will add the
ble_controller/include to the list of include directories, but not link
the actual controller.

Signed-off-by: Thomas Stenersen <thomas.stenersen@nordicsemi.no>