ble_controller: Add config to default to BT_LL_NRFXLIB
If CONFIG_BT_LL_NRFXLIB_DEFAULT=y, the BT_LL_NRFXLIB controller will be
selected by default if the platform supports it. You can set this option
in a generic prj.conf file and it will work when you're building for
nrf5340_cpuapp as well, and not have to create custom configuration
files for all the platforms.

Signed-off-by: Thomas Stenersen <thomas.stenersen@nordicsemi.no>