Bluetooth: controller: Add ECDH support in the softdevice controller
Add ECDH support in the softdevice controller, using either tinycrypt
or the oberon crypto library.
Supporting the ECDH in the controller allows the ECDH calculations
to be done on the controller side in a split host-controller build.
Using the oberon crypto library has better performance, stack and
flash usage compared to tinycrypt.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>