nrf_security: cmake: archive combining is now a custom CMake target
Archive combining is now a custom CMake target instead of a CMake
library target and an interface lib is defined for linking to allow
linking to the combined library.

This has a couple of benefits:
- empty_file.c is no longer needed, this cleans up the code
- The combined archive is a dedicated target, upon which other targets
  can depend
- Removed POST_BUILD step, which further cleanup the code.

This fixes a particular issue where repackaging of libraries didn't
work correctly which were observed in SES-NE.

This commit ensures that build works correctly on all platforms and
with SES / Ninja / Make.

Fixes: NCSDK-6989

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>