compliancecheck: using KconfigBasic compliance check instead of Kconfig
Using `compliancecheck` in sdk-nrf generates >500 undefined Kconfig
symbol references. Many of those are false warnings, but to allow time
for proper cleanup, then this commit resorts to only do KconfigBasic
check.

This means that undefined Kconfig symbols in the Kconfig tree are
reported but not outside the tree, such as source code, rst files, conf
files.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>