zboss: Simplify ZBOSS CMake targets
After adding a dependency on the nrfxlib module inside ZBOSS plaform,
there is no need to define zboss-extensions CMake target.

Due to the Kconfig option override rules, the library type symbol has to
be dependent on the existance of ZBOSS sources in the west project.
After adding the dependency - the overriding Kconfig will be processed
after the one that intorduces option with defaults.

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>