nrf_security: cmake: refactor library re-archiving
Fixes: NCSDK-10008

SES-NE has a limitation on how long a POST_BUILD command from the
build.ninja it can handle.

When that length is exceeded then post build commands are truncated in
the middle and thus causing unexpected behavior.

This commit creates a <backend>_extract build target which can then be
set as a depedency to the static library. As the extract target is a
dependency it will be executed prior to the build of the static library.
This ensures that the post build step of adding objects to the static
library can now be safely done as the post build step is now within the
limits of SES-NE.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>