nrf_security: cmake: using archive script when archiving a folder
Fixes: NCSDK-10008

When archiving extracted objects from imported nRF security libraries
then a globbing expression in the form of: `ar q <library> <folder>/*`
is used.

Unfortunately, SES-NE treat the `*` as a filename and not a globbing
expression which results in `ar` to fail as there is no file named `*`.

This is handled by introducing archive_folder.cmake script which takes
the folder as an argument and then globs for all files in that folder.

The list of files is then passed directly to the `ar` command including
the library, and thus no globbing is needed directly on `ar`.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>