nrf_security: symbol rename script now passes settings through script
The symbol rename archive script used a generator expression to pass
all objects to the symbol_rename_archive_script.cmake among other
arguments.

When there are many objects in the library, then the command execution
string becomes very long.

On Windows, long execution commands are
wrapped inside a bat-file.

This might look like:
...\post-build.bat 7eb3ae9870a59d55

Resulting in SES-NE to try to execute the post-build.bat including the
following SHA as a single command, which fails.

This is fixed by generating a CMake file with required arguments to
symbol_rename_archive_script.cmake, and the pass this settings file as
argument to the script.

Fixes: NCSDK-10510

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>