nrf_802154: fix build warning for nrf5340 APP
The missleading warning is displayed: "This combination of SoC
and floating point ABI is not supported by the nrf_802154_sl lib."
for the case where nrf_802154_sl lib is not linked (it happens for
target: nrf5340dk_nrf5340_cpuapp when the radio serialization is
used instead).
This commit fixes it by exluding "sl" directory when the option
CONFIG_NRF_802154_RADIO_DRIVER is not active.

Signed-off-by: Lukasz Maciejonczyk <lukasz.maciejonczyk@nordicsemi.no>