checkpatch: add `--exclude-dir nrf_modem/include`
nrf_socket.h must provide type definitions, which make checkpatch
emit a warning if the type definition is not in the typedefsfile.

Unfortunately that file is in zephyr/scripts/checkpatch/
which means that to add an entry there, a patch must be
merged to sdk-zephyr first, and the sdk-nrf manifest
must be updated accordingly (since the GitHub action
in this repository simply checks out sdk-nrf/main).

For now, simply add the whole nrf_modem/include folder
to the --exclude-dir list in checkpatch.conf in this repo.

Eventually, the typedefsfile in sdk-zephyr could be updated
to ignore the necessary type definitions in nrf_modem headers,
if preferred.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>