nrf_security: Remove NRF_SECURITY_RNG
This removes NRF_SECURITY_RNG since we don't
have a clear use case for it and it can
be confusing for users. We now support the
Zephyr RNG model for TF-M  builds.
The NRF_SECURITY_RNG was also causing issues
when used with the PSA RNG from Zephyr which
we use for the TF-M builds now.

Ref: NCSDK-14509

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>