nrf_security: Add mbedtls_hkdf function to PSA Crypto config
Add legacy function mbedtls_hkdf to PSA Crypto configuration.
This is needed by the TF-M builtin key loader.

tfm_builtin_key_loader.c:
FIXME: this should be moved to using the PSA APIs once key derivation is
implemented in the PSA driver wrapper. Using the external PSA apis
directly creates a keyslot and we'd need to read the data from it and
then destroy it, so isn't ideal. In order to avoid infinite recursion,
it'll be necessary to add a special case (probably if owner == 0) to
make sure the new PSA derivation request doesn't end up back here.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>