nrf_security: Allow import key from PSA core
This allows the psa_import_key to use the
implementation for symmetric key types.
This is useful because we can later remove the
import key from the different drivers and reduce
the code size and maintenance of the drivers. The
effect that this will have in current builds is
very minimal.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>