nrf_security: Use export_public_key from PSA core
This makes sure that we always call the function
psa_export_public_key_internal when no crypto accelerator
export_public_key functionality is available. This is needed
because our CC3XX library does not support export_public_key
when the key itself is a public key.
The psa_export_public_key_internal is configurable and
thus the code size change should be minimal.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>