nrf_security: Remove double dependency on PSA crypto driver configs
Remove double dependency on PSA crypto driver configurations.
The was in if check for rsource-ing the file, then an additional
if inside the file itself.

Also move the menuconfig entry inside the driver Kconfig file.
This makes the file self-contained, and also easier to add possible
driver configurations that would always be enabled regardless of the
main configuration.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>