nrf_security: Fix interoperability issue between Oberon PSA and CC3XX
Fix interoperability issue between Oberon PSA core and CC3XX PSA crypto
driver.
The driver does not reset the operation context on succes as the PSA
core expects it too. With Mbed TLS builtin PSA core this issue is not
seen because the core always calls abort to the driver.

NCSDK-21377

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>