nrf_security: cmake: Remove and update references to nrf_security
Remove references to nrf_security.

In Kconfig we source nrf from nrfxlib such that the sources are still
processed in the same order as before. Changing the order of Kconfig
sourcing would require adding _OVERRIDE symbols for each Kconfig
option and we would like to do this after the move to unfreeze
nrf_security.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>