mpsl: Fix problems with building separate MPSL and FEM libraries
Due to the fact, that the FEM libraries are dependent on the MPSL
library, some applications fail to build if the MPSL library
is passed first to the linker (as the declared functions are
used later, and are deleted in the optimization process).

This commit makes sure that the FEM libraries are passed first,
so that the MPSL functions are referenced before they are defined,
so that the linker does not delete them.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>