doc: crypto: Refer to nrf91 instead of nrf9160
Update crypto docs to refer to nrf91 instead of nrf9160 for
portability.

Also, update one place where nrf53 was omitted.

Only the most recent changelog is updated as changelogs do not really
need to be updated. We only update the changelogs to make it easier to
copy-paste the changelog on the next release.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>