nrf_rpc: Init all transports before sending group init packets
This commit postpones sending group init packets to
after all transports have had a chance to initialize.

This ensures that the initialization of one transport don't have
to wait for the group init packets being sent on another transport.
This could occur if the send function of a transport would block.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>