mpsl: Use kconfig to determine library path
Instead of using SOC and ARCH configs to determine the library
paths, use mpsl-specific kconfigs. We will then be able to change
the paths independently of which value various upstream configs
have.

This also makes it simpler to use the same library for different socs,
or different libraries for the same cpu core variant.

Signed-off-by: Herman Berget <herman.berget@nordicsemi.no>