nrf_rpc: Two-way handshake initialization
Implement two-way handshake nRF RPC initialization:
When an initialization packet with the unknown destination
group ID is received, reply to that with an initialization
packet with the right destination group ID.

This makes it possible to re-establish the link even if
either peer has rebooted, or has booted with a delay.

Also, allow to configure infinite group init timeout.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>