nrf_wifi: Implement RPU recovery trigger
When a watchdog interrupt is received, verify the conditions and see if
a RPU recovery needs to be performed or not.

For now we just check if host as asserted WAKEUP_NOW or not, and if not
then watchdog interrupt is unexpected and needs a recovery.

HAL processes the IRQ and decides if it needs a recovery or not, and
FMAC relays the recovery for the native code to implement the recovery
based on the OS.

Implements SHEL-2726.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>