cmake: nrfxlib_calculate_lib_path NRF54H20_ENGB_CPURAD
One library path is returned by nrfxlib_calculate_lib_path for both
CONFIG_SOC_NRF54H20_CPURAD and CONFIG_SOC_NRF54H20_ENGB_CPURAD.
The libraries for both version of the SoC are expected to be the same.

In case any user of this function needs a separate version
(not recommended) of the library for CONFIG_SOC_NRF54L15_ENGA_CPUAPP,
the result returned by this function should be updated by the
calling cmake code.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>