nrf_802154: fix add missing select NRFX_GPPI
When CONFIG_NRF_802154_SL=y on the nRF52 and nRF53 Series the
sdk-nrf uses the nrf_802154_platform_sl_lptimer_zephyr.c file.
This file explicitly depends on NRFX_GPPI api.

The NRFX_GPPI was selected indirectly by other modules, but if
the other modules are turned off, the NRFX_GPPI was not selected.

This commit fixes that. The CONFIG_NRFX_GPPI is selected When
nRF 802.15.4 Radio Driver with closed-source SL is used on
nRF52 and nRF53 series.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>