mpsl: Improve GRTC syscounter performance
Mandate NRF_GRTC_TIMER_AUTO_KEEP_ALIVE for platforms that have the GRTC.
This improves the worst case latency for reading the syscounter value.

NRF_GRTC_TIMER_AUTO_KEEP_ALIVE together with the grace period of the
GRTC should be sufficient to guarantee a low-latency for reading the
SYSCOUNTER where the latency actually matters.

This is needed to provide the performance that MPSL and the SoftDevice
controller expect. Without NRF_GRTC_TIMER_AUTO_KEEP_ALIVE it may take up
to around 100us to read the GRTC when it is in the inactive state.

Signed-off-by: Johan Stridkvist <johan.stridkvist@nordicsemi.no>