rpc: Implement possibility to stop and resume all the RPC communication
Implemented stop function that immedtaietly releases all threads waiting
for RPC response and automatically fails all new calls.
Implemented resume function that restores normal RPC communication.
This is intended to be used for recovery process when the other side
needs to be reset but there are calls waiting for the response.
Additionally it is possible to provide callback handler that could be
used to do the custom cleanup, for example clear the callback table.

Signed-off-by: Marek Porwisz <marek.porwisz@nordicsemi.no>