nrf_rpc: fix initialization corner case
When initializing multiple groups that share the same
transport, it may happen that a device receives an init
packet for a group whose internal members have not been
initialized yet, leading to the bus fault.

Make sure that internal members of all groups have been
initialized before ANY transport is started to prevent
that.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>