nrf_rpc: add Kconfig to configure default group flags
By default, groups defined with NRF_RPC_GROUP_DEFINE()
are initiators (send an init packet to the peer) and wait
for initialization (until they receive an init packet from
the peer).

Make it possible to override these settings with Kconfig
so that one can decide which device is an initiator and
which is a responder.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>