nfc: remove depends on HAS_HW_NRF_NFCT due to sdk-nrf-bm
The sdk-nrfxlib should be platform agnostic. In sdk-nrf-bm the
device tree is not used. The HAS_HW_NRF_NFCT Kconfig symbols are
not provided despite having a capable hardware.
Because of that the dependency on HAS_HW_NRF_NFCT symbol provided by
only one platform (full nRF Connect SDK) is removed.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>